requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.8.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
namex==0.0.8
numpy==1.26.4
optree==0.14.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
six==1.17.0
tensorflow-cpu==2.17.1  
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
